requirements.txt
# Face Recognition Dataset Project Requirements
# Project ID: 22 - Face Recognition Dataset
# RSK World - https://rskworld.in/
# Founder: Molla Samser
# Designer & Tester: Rima Khatun
# Email: help@rskworld.in
# Phone: +91 93305 39277
# Year: 2026

numpy>=1.21.0
opencv-python>=4.5.0
face-recognition>=1.3.0
Pillow>=8.0.0
scikit-learn>=0.24.0
matplotlib>=3.3.0
pandas>=1.3.0
flask>=2.0.0
requests>=2.25.0
dlib>=19.22.0

